Changes in the finite domain constraint programming support. - Renamed fd propagators builtins file. - Created file to host auxiliary functions of fd propagators builtins.
